ImageBufferIOSurfaceBackend should get and put data in WP when in mapped mode https://bugs.webkit.org/show_bug.cgi?id=252987 rdar://105972277
Reviewed by NOBODY (OOPS!).

In non-IOKit blocking mode, WP has the access to the IOSurface.

To do getImageData(), it is redundant to ask for the data from
GPUP. The more performant approach is to flush the rendering and lock the
IOSurface for read.

To do putImageData(), it is currently mostly not performant to send
the data to GPUP. When the GPUP gets the data, it will flush the
rendering and lock the IOSurface for write. This causes the synchronization
with the rasterization and GPU.
Since currently we have the synchronization, just use the same approach
from WP. Send a command to invalidate the image caches and wait for that
command to be flushed. When this flush completes, nothing in GPUP will
access the IOSurface. Then lock the IOSurface fsor write and put the
data there.

* Source/WebKit/WebProcess/GPU/graphics/RemoteImageBufferProxy.cpp:
(WebKit::RemoteImageBufferProxy::getPixelBuffer const):
(WebKit::RemoteImageBufferProxy::putPixelBuffer):