Feedback for Matt's 1c8f7e6ccc82.
Matt has since force-pushed 9248afe7104c but the differences are not
changing this feedback.

The branch compiles but I don't know at all how it runs. Probably
does not run. Does not address any differences between
1c8f7e6ccc82 and 9248afe7104c.

The code was in such a shape that I did not have the time to address
the identifier exchange problem at all. As such, the design is based on
the current, flawed but somewhat working for this case, design
where UI process allocates a remote snapshot identifier and WCPs
write renderings to <global snapshot identifier, global frame identifier>
entries.

Until the syntactic changes land, there's no real point in investing
in a better identifier handling mechanism.

Review comments:

** Do not add DisplayItems::DrawRemoteFrame.

"Drawing remote frame" is not part of the display list or graphics
context domain of functionality. It does not make sense for any
display list.

Instead, add "draw placeholder item" which can draw arbitrary graphics
context commands. This concept stays within graphics context domain.

This way you do not need the redundant resolve dependencies -phase.
It is redundant to go through all the display list items and replace
contents.

** Do not add RemoteDisplayListRecorder::DrawRemoteFrame.

Not all RemoteDisplayListRecorders know what to do with DrawRemoteFrame
command. What happens if non-Snapshot RemoteDisplayListRecorder
encounters such a command?

Instead, add RemoteSnapshotRecorder. RemoteSnapshotRecorder is capable
of resolving frame snapshots for the particular snapshot.
Add snapshot specific functionality:
RemoteSnapshotRecorder::DrawSnapshotFrame with the semantics that
the frame snapshot will be resolved once the Snapshot is fully complete.
Fully complete snapshots are such that have all their frame snapshot
references resolved.

** Do not add snapshot-specific recorder to
RemoteRenderingBackend::m_remoteDisplayListRecorder list.

This creates the vulnerability where WCP can first create snapshot recorder
and then sink it into normal display list and then use that display list as
a source to normal bitmap draw. This way WCP can extract subframe bitmaps.

** Do not use eager cancellation of the remote Snapshot upon error

The autovivification of Snapshot needs that all the snapshot creation
events have been waited upon before the snapshot can be cancelled.
Otherwise we risk leaks, since some snapshot creation tasks
might not have started in GPUP before the snapshot cancel task already
complete in GPUP.

Consider
(WCP-WCP sends below go via UI of course):
In WCP Main:
    - sends WCP Sub1: drawFrameToSnapshot
    - sends WCP Sub2: drawFrameToSnapshot
WCP Sub1 crashes

In WCP Main:
    - Notice callback for Sub1 fails
    - Fail the UI callback

UI:
    - Notice WCP main drawToSnapshot callback fails
    - Send cancellation
    * In Matt's, it is retrievePDFSnapshot success=false

GPUP:
    - retrievePDFSnapshot success=false
    -> remove the snapshot
    - continue processing messages
    - notice Sub2 drawFrameToSnapshot
        -> create the snapshot again, and leak it.

** Do not name the snapshot as "SnapshotCompositor"

The compositing of the snapshot is very ephemeral detail of the snapshot
and none of the vocabularity related to manipulating the snapshot
refers to that. The domain object is in the code is "Snapshot" that
can be pointed to with an identifier.
Suggestion: use RemoteSnapshot.

** Do not implement the snapshot as inner class of GPUProcess

The class is of substantial complexity and deserves all the
intellectual space that it can afford, including its of files.

** Failure modes with adversary WCP are not considered.

Suggestion: consider the obvious ones.

Other suggestions in the patch here:

* Add Snapshot recording functionality as refcounted RemoteSnapshot

* Move WebCore::SnapshotIdentifier as WebKit::RemoteSnapshotIdentifier

* Model the drawing loop as such:
    - WCPs create RemoteSnapshotRecorders for a particular snapshot
    - WCPS draw via the recorders and sink the renderings to slots
    identified by FrameID for their own snapshot.
    - WCPS refer to other frame renderings, within their own snapshot
    - Main WCP returns the main content size to UI as the result of
    its renderings
    - UI sinks the snapshot to PDF with the content size and the initial
    frame id

The last two facts try to overcome the rendundant complexity that
comes from the separate handling of mainframe and subframes
in the existing code. E.g. it does not make sense that the mainframe
display list is accompanied with the size but the subframe is not.
The size is used to size the PDF.

* ifdefs:
    - "Render to remote snapshot" is common to all ports
    - "Convert remote snapshot to PDF" is Cocoa feature